Fix `stop` of socket server
The socket server currently does not actually shutdown the server
when calling shutdown.
In order to do so, the `close` method must be called on the server.
Thus the server was moved into a private member
to be able to call `close` on `stop`.
This also fixes the test launch configs and adds a test.
Furthermore the overriden methods in SocketServerLauncher
are not public anymore in order to have a clean api.

Contributed on behalf of STMicroelectronics.