GLSP-1109: Remove dependency to fs-extra (#56)
Use 'fs' instead of 'fs-extra' in GLSP server node as the additional capatibilities of
'fs-extra' are not used. Also fixes the issue that the dependency to 'fs-extra' was not
declared in the 'package.json' which might fail consumer builds.

Fixes https://github.com/eclipse-glsp/glsp/issues/1109

Contributed on behalf of STMicroelectronics.